+++ pkg/platform/kube/crd/BUILD
@@ -22,7 +22,6 @@ go_library(
         "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime/serializer:go_default_library",
         "@io_k8s_apimachinery//pkg/util/wait:go_default_library",
-        "@io_k8s_client_go//kubernetes:go_default_library",
         "@io_k8s_client_go//plugin/pkg/client/auth/gcp:go_default_library",
         "@io_k8s_client_go//plugin/pkg/client/auth/oidc:go_default_library",
         "@io_k8s_client_go//rest:go_default_library",
@@ -49,5 +48,6 @@ go_test(
         "//pkg/model/config:go_default_library",
         "//pkg/testing/mock:go_default_library",
         "//pkg/testing/util:go_default_library",
+        "@io_k8s_client_go//kubernetes:go_default_library",
     ],
 )
