+++ pkg/platform/kube/crd/client.go
@@ -18,6 +18,7 @@ package crd
 
 import (
 	"fmt"
+	"os"
 	"time"
 
 	"github.com/golang/glog"
@@ -30,12 +31,9 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/kubernetes"
 	// import GKE cluster authentication plugin
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	// import OIDC cluster authentication plugin, e.g. for Tectonic
-	"os"
-
 	_ "k8s.io/client-go/plugin/pkg/client/auth/oidc"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -89,21 +87,6 @@ func resolveConfig(kubeconfig string) (string, error) {
 		}
 	}
 	return kubeconfig, nil
-}
-
-func createInterface(kubeconfig string) (*rest.Config, kubernetes.Interface, error) {
-	kube, err := resolveConfig(kubeconfig)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	config, err := clientcmd.BuildConfigFromFlags("", kube)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	client, err := kubernetes.NewForConfig(config)
-	return config, client, err
 }
 
 // CreateRESTConfig for cluster API server, pass empty config file for in-cluster
@@ -195,8 +178,8 @@ func (cl *Client) RegisterResources() error {
 				Scope:   apiextensionsv1beta1.NamespaceScoped,
 				Names: apiextensionsv1beta1.CustomResourceDefinitionNames{
 					Singular: schema.Type,
-					Plural: schema.Plural,
-					Kind:   kabobCaseToCamelCase(schema.Type),
+					Plural:   schema.Plural,
+					Kind:     kabobCaseToCamelCase(schema.Type),
 				},
 			},
 		}
