+++ pkg/testing/mock/config.go
@@ -23,9 +23,9 @@ import (
 
 	"github.com/golang/protobuf/proto"
 
+	brokerconfig "istio.io/api/broker/v1/config"
 	"istio.io/broker/pkg/model/config"
 	testproto "istio.io/broker/pkg/testing/mock/proto"
-	brokerconfig "istio.io/api/broker/v1/config"
 )
 
 var (
